requirements.txt
# Hub-Bridging Validation Framework Dependencies
# Core scientific computing
numpy>=1.24
scipy>=1.10
pandas>=2.0

# Network analysis
networkx>=3.0
python-louvain>=0.16
leidenalg>=0.10
igraph>=0.10

# Statistical analysis
scikit-learn>=1.3
powerlaw>=1.5
statsmodels>=0.14

# Visualization
matplotlib>=3.7
seaborn>=0.12

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.4
pytest-cov>=4.1

# Jupyter notebooks
jupyter>=1.0
ipykernel>=6.25

# Development tools
black>=23.0
isort>=5.12
mypy>=1.5

# Documentation
sphinx>=7.0
sphinx-rtd-theme>=1.3
